Move files to `modules` dir
